Functions and utilities to export hiplot output

Functions and utilities to export hiplot output

Usage

export_single(x)

export_pptx(x, outfn, width, height)

export_plotly(x, outfn)

export_directory(
  outdir = "output",
  pdf_tb_subdir = c("output", "output/results"),
  recursive = c(FALSE, TRUE),
  sheetname_replace = "_BayesianNMF|legacy_fitting_|_fitting|.csv"
)

export_htmlwidget(x)

html2pdf(inhtml, outpdf)

pdf2image(pdf, i = "")

pdfs2image(pdfs)

merge_pdfs(pdfs, outpdf)

split_pdfs(pdfs, outdir)

import_images_to_pdf(pdfs)

read_data(path, format = c("csv", "excel", "rds", "rdata"), header = TRUE, ...)

parse_file_link(link)

Arguments

x

a plot object to export.

outfn

output file path.

width

width of output pptx.

height

height of output pptx.

outdir

output directory.

pdf_tb_subdir

directory path storing pdf (figure) files.

recursive

same path as pdf_tb_subdir to specify if run with recursive mode.

sheetname_replace

all .csv files are re-exported into a excel file, if necessary, use this option remove some string as sheet names.

inhtml

input html file.

outpdf

output pdf file.

pdf

input pdf file.

i

a postfix added to the output file name.

pdfs

input pdf files.

path

the file path to data.

format

import data format.

header

if TRUE, tread input having a header.

...

other parameters passing to data.table::fread when use this function to import inside.

link

the file link

Functions

export_single: Export a single ggplot object

export_pptx: Export a single ggplot object to PPTx

export_plotly: Export a single ggplot object to plotly

export_directory: Export hiplot plugin output directory to a zipped file

export_htmlwidget: Export a single ggplot object to html file

html2pdf: Convert html to pdf

pdf2image: Convert pdf to images

pdfs2image: Convert multiple pdfs to images

merge_pdfs: Merge multiple pdf files into one

split_pdfs: Split multiple page pdfs to a directory

import_images_to_pdf: Export non-pdf images as pdfs

read_data: Read common format data into data frame

parse_file_link: Parse the true file path from link

Examples

# parse_file_link("hiplot.org/path=xxx") TODO