Prestored settings and functions related to ggplot2 or plots

Prestored settings and functions related to ggplot2 or plots

Usage

cb_palette

ggsci_palette_names

ggsci_palette_length

brewer_palette_names

prism_palette_names

graf_palette_names

ggplot2_themes

ggplot2_themes2

choose_ggplot_theme(p, theme = "theme_bw")

return_ggplot_theme(theme = "theme_bw")

position_type(label = c("dodge", "fill", "stack"))

set_palette_theme(p, conf = globs_get("conf"))

set_complex_general_theme(p, conf = globs_get("conf"))

return_hiplot_palette(u, custom = NULL)

return_hiplot_palette_color(u, custom = NULL)

get_hiplot_color(u, n = -1, custom = NULL)

gg_color_default(n)

brewer_pal_update(n, panel = "Set3")

transparentColor(col, alpha = 200)

custom_color_filter(cols)

alter_fun(x, y, w, h, v)

col_fun_cont(
  x,
  cols = c("#196ABD", "#3399FF", "#3399FF", "#f4f4f4", "#f4f4f4", "#f4f4f4", "#FF3333",
    "#FF3333", "#C20B01")
)

col_tag

mut_cols

draw_map(rds, keyname, filter_names = NULL)

Format

An object of class character of length 37.

An object of class character of length 16.

An object of class numeric of length 16.

An object of class character of length 9.

An object of class character of length 65.

An object of class character of length 3.

An object of class character of length 50.

An object of class list of length 6.

An object of class character of length 2.

An object of class character of length 30.

Arguments

p

a ggplot object.

theme

a string specifying theme.

label

a position label, one of 'dodge', 'fill' and 'stack'.

conf

a conf object, see conf.

u

one of a predefined palette name. See ggsci_palette_names, brewer_palette_names, prism_palette_names, etc.

custom

a user specified custom color vector.

n

number of colors.

panel

defines the color panel you want

col

a color.

alpha

a alpha value.

cols

colors.

x, y, w, h, v

a set of parameters passing to grid::grid.rect.

rds

a file storing in the format of .RDS.

keyname

a keyname to plot.

filter_names

used for filtering.

Functions

cb_palette: A vector of colors

ggsci_palette_names: A vector of ggsci palette names

ggsci_palette_length: Maximum color number provided by palette of ggsci_palette_names

brewer_palette_names: A vector of RColorBrewer palette names

prism_palette_names: A vector of prism palette names

graf_palette_names: A vector of graf palette names

ggplot2_themes: A vector of ggplot2 themes

ggplot2_themes2: A list of ggplot2 themes

choose_ggplot_theme: Set ggplot2 theme for a ggplot object

return_ggplot_theme: Return ggplot2 theme based on string

position_type: Return position type

set_palette_theme: Set palette and theme from conf for a ggplot object

set_complex_general_theme: Set complex general theme from conf for a ggplot object

return_hiplot_palette: Return ggplot2 scale of fill palette with predefined or custom colors

return_hiplot_palette_color: Return ggplot2 scale of color palette with predefined or custom colors, similar to return_hiplot_palette

get_hiplot_color: Get colors from a color palette

gg_color_default: A color generator for default colors in ggplot

brewer_pal_update: An update version of brew.pal function

transparentColor: Transparent colors

custom_color_filter: A custom color filter

alter_fun: A alter_fun used in ComplexHeatmap

col_fun_cont: A color function used in ComplexHeatmap

col_tag: A col_tag used in ComplexHeatmap (TODO: make sure)

mut_cols: A mapping between mutation type and colors used in ComplexHeatmap (TODO: make sure)

draw_map: Draw a Map

Examples

transparentColor("red")
#> [1] "#FF0000C8"
custom_color_filter("AAAAAA, CCCCCC")
#> [1] "#AAAAAA" "#CCCCCC"